create class Session
