+++ schedule/schedule.py
@@ -1,132 +1,14 @@
 import datetime
 import time
 from threading import Thread
-import abc
-from typing import Optional
 
-import numpy as np
-from sqlalchemy import select, func, or_
+from sqlalchemy import select
 
 from models import db_session
 from models.db_session import create_session
-from models.questions import Question, QuestionAnswer, AnswerState
-from models.users import Person, PersonGroup, PersonGroupAssociation
+from models.questions import AnswerState
+from models.users import Person
 from tools import Settings, WeekDays
-
-
-class GeneratorInterface(abc.ABC):
-    @abc.abstractmethod
-    def next_question(self, person: Person) -> QuestionAnswer:
-        pass
-
-    @staticmethod
-    def _get_planned(db, person: Person) -> Optional[QuestionAnswer]:
-        return db.scalars(select(QuestionAnswer).
-                          where(QuestionAnswer.person_id == person.id,
-                                QuestionAnswer.ask_time <= datetime.datetime.now(),
-                                QuestionAnswer.state == AnswerState.NOT_ANSWERED).
-                          order_by(QuestionAnswer.ask_time)).first()
-
-    @staticmethod
-    def _get_person_questions(db, person: Person) -> list[Question]:
-        return db.scalars(select(Question).
-                          join(Question.groups).
-                          where(PersonGroup.id.in_(pg.id for pg in person.groups)).
-                          group_by(Question.id)).all()
-
-
-class SimpleRandomGenerator(GeneratorInterface):
-    def next_question(self, person: Person) -> QuestionAnswer:
-        with db_session.create_session() as db:
-            planned = self._get_planned(db, person)
-            if planned is not None:
-                return planned
-
-            person_questions = self._get_person_questions(db, person)
-
-            question = np.random.choice(person_questions)
-
-            cur_answer = QuestionAnswer(question_id=question.id,
-                                        person_id=person.id,
-                                        ask_time=datetime.datetime.now(),
-                                        state=AnswerState.NOT_ANSWERED)
-
-            db.add(cur_answer)
-            db.commit()
-
-        return cur_answer
-
-
-class StatRandomGenerator(GeneratorInterface):
-    def next_question(self, person: Person) -> QuestionAnswer:
-        with db_session.create_session() as db:
-            planned = self._get_planned(db, person)
-            if planned is not None:
-                return planned
-
-            person_questions = self._get_person_questions(db, person)
-            probabilities = np.zeros(len(person_questions))
-
-            for i, question in enumerate(person_questions):
-                question: Question
-                correct_count = db.scalar(select(func.count(QuestionAnswer.id)).
-                                          join(QuestionAnswer.question).
-                                          where(QuestionAnswer.person_id == person.id,
-                                                QuestionAnswer.question_id == question.id,
-                                                QuestionAnswer.person_answer == Question.answer))
-
-                if correct_count:
-                    last_correct_or_ignored = db.scalar(select(QuestionAnswer).
-                                                        join(QuestionAnswer.question).
-                                                        where(QuestionAnswer.person_id == person.id,
-                                                              QuestionAnswer.question_id == question.id,
-                                                              or_(QuestionAnswer.person_answer == Question.answer,
-                                                                  QuestionAnswer.state != AnswerState.NOT_ANSWERED)).
-                                                        order_by(QuestionAnswer.ask_time.desc()))
-
-                    first_answer = db.scalar(select(QuestionAnswer).
-                                             where(QuestionAnswer.person_id == person.id,
-                                                   QuestionAnswer.question_id == question.id))
-
-                    answers_after_first = db.scalar(select(func.count(QuestionAnswer.id)).
-                                                    where(QuestionAnswer.person_id == person.id,
-                                                          QuestionAnswer.ask_time > first_answer.ask_time,
-                                                          QuestionAnswer.state != AnswerState.NOT_ANSWERED)) + 1
-
-                    max_target_level = db.scalar(select(func.max(PersonGroupAssociation.target_level)).
-                                                 where(PersonGroupAssociation.person_id == person.id,
-                                                       PersonGroupAssociation.group_id.in_(
-                                                           q.id for q in question.groups)))
-
-                    p = (datetime.datetime.now() - last_correct_or_ignored.ask_time).total_seconds() / correct_count
-                    p *= np.abs(np.cos(np.pi * np.log2(answers_after_first + 4))) ** (
-                            ((answers_after_first + 4) ** 2) / 20)  # planning questions
-                    p *= np.e ** (-0.5 * (max_target_level - question.level) ** 2)  # normal by level
-
-                    probabilities[i] = p
-                else:
-                    probabilities[i] = None
-
-            with_val = list(filter(lambda x: not np.isnan(x), probabilities))
-            without_val_count = len(person_questions) - len(with_val)
-
-            increased_avg = (sum(with_val) + without_val_count * max(with_val)) / len(person_questions)
-
-            probabilities[np.isnan(probabilities)] = increased_avg
-
-            probabilities /= sum(probabilities)
-
-            question = np.random.choice(person_questions, p=probabilities)
-
-            cur_answer = QuestionAnswer(question_id=question.id,
-                                        person_id=person.id,
-                                        ask_time=datetime.datetime.now(),
-                                        state=AnswerState.NOT_ANSWERED)
-
-            db.add(cur_answer)
-            db.commit()
-
-        return cur_answer
 
 
 class Schedule(Thread):
@@ -142,8 +24,6 @@ class Schedule(Thread):
         Settings().add_update_handler(self.from_settings)
 
         self.previous_call = None
-
-        self.generator: GeneratorInterface = StatRandomGenerator()
 
     def from_settings(self):
         self._every = Settings()['time_period']
@@ -172,22 +52,7 @@ class Schedule(Thread):
             time.sleep(1)
 
     def task(self):
-        question_for_person = []
         with db_session.create_session() as db:
             persons = db.scalars(select(Person).where(Person.is_paused.is_(False)))
             for person in persons:
-                answer = self.generator.next_question(person)
-                question_for_person.append((person, answer))
-
-        self._send_to_people(question_for_person)
-
-    def _send_to_people(self, question_to_person):
-        with create_session() as db:
-            for person, answer in question_to_person:
-                answer = db.merge(answer)
-                person = db.merge(person)
-
-                answer.state = AnswerState.TRANSFERRED
-                db.commit()
-
-                self._callback(person, answer)  # Send a question to a person
+                self._callback(person)
